remove unused expressions

+++ src/main/java/br/com/cpa/spring/modules/user/patient/services/DeletePatientService.java
@@ -2,7 +2,9 @@ package br.com.cpa.spring.modules.user.patient.services;
 
 import br.com.cpa.spring.modules.user.patient.PatientRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.server.ResponseStatusException;
 
 import java.util.UUID;
 
@@ -14,6 +16,9 @@ public class DeletePatientService {
     private PatientRepository patientRepository;
 
     public void execute(Long id) {
+        if (!patientRepository.existsById(id)) {
+            throw new ResponseStatusException(HttpStatus.NOT_FOUND, "User not found");
+        }
         patientRepository.deleteById(id);
     }
 }
